opponent-detector/requirements.txt
# Labeling
label-studio>=1.11

# Training
ultralytics>=8.3
opencv-python
pyyaml
matplotlib
tqdm
ipykernel